Moved Context and its impls to a separate file
